moved monkey patching to an auto-use fixture
